Import Upstream version 1.41

+++ tests/tst_nfkc.c
@@ -1,5 +1,5 @@
 /* tst_nfkc.c --- Self tests for stringprep_utf8_nfkc_normalize().
- * Copyright (C) 2002-2016 Simon Josefsson
+ * Copyright (C) 2002-2022 Simon Josefsson
  *
  * This file is part of GNU Libidn.
  *
@@ -14,7 +14,7 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  *
  */
 
@@ -68,18 +68,18 @@ void
 doit (void)
 {
   char *out;
-  size_t i;
+  unsigned i;
 
   for (i = 0; i < sizeof (nfkc) / sizeof (nfkc[0]); i++)
     {
       if (debug)
-	printf ("NFKC entry %ld\n", i);
+	printf ("NFKC entry %u\n", i);
 
       out = stringprep_utf8_nfkc_normalize (nfkc[i].in,
 					    (ssize_t) strlen (nfkc[i].in));
       if (out == NULL)
 	{
-	  fail ("NFKC entry %ld failed fatally\n", i);
+	  fail ("NFKC entry %u failed fatally\n", i);
 	  continue;
 	}
 
@@ -114,7 +114,7 @@ doit (void)
       if (strlen (nfkc[i].out) != strlen (out) ||
 	  memcmp (nfkc[i].out, out, strlen (out)) != 0)
 	{
-	  fail ("NFKC entry %ld failed\n", i);
+	  fail ("NFKC entry %u failed\n", i);
 	  if (debug)
 	    printf ("ERROR\n");
 	}
