+++ tests/tst_tld.c
@@ -1,5 +1,5 @@
 /* tst_tld.c --- Self tests for tld_*().
- * Copyright (C) 2004-2016 Simon Josefsson
+ * Copyright (C) 2004-2022 Simon Josefsson
  *
  * This file is part of GNU Libidn.
  *
@@ -14,7 +14,7 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  *
  */
 
@@ -62,25 +62,23 @@ static const struct tld tld[] = {
    3}
 };
 
-static const Tld_table _tld_fr_override =
-  {
-    "fr",
-    "2.0",
-    0,
-    NULL
-  };
+static const Tld_table _tld_fr_override = {
+  "fr",
+  "2.0",
+  0,
+  NULL
+};
 
 /* Main array */
-const Tld_table * my_tld_tables[] =
-  {
-    &_tld_fr_override,
-    NULL
-  };
+const Tld_table *my_tld_tables[] = {
+  &_tld_fr_override,
+  NULL
+};
 
 void
 doit (void)
 {
-  size_t i;
+  unsigned i;
   const Tld_table *tldtable;
   char *out;
   size_t errpos;
@@ -206,7 +204,7 @@ doit (void)
   for (i = 0; i < sizeof (tld) / sizeof (tld[0]); i++)
     {
       if (debug)
-	printf ("TLD entry %ld: %s\n", i, tld[i].name);
+	printf ("TLD entry %u: %s\n", i, tld[i].name);
 
       if (debug)
 	{
@@ -217,7 +215,7 @@ doit (void)
       tldtable = tld_default_table (tld[i].tld, NULL);
       if (tldtable == NULL)
 	{
-	  fail ("TLD entry %ld tld_get_table (%s)\n", i, tld[i].tld);
+	  fail ("TLD entry %u tld_get_table (%s)\n", i, tld[i].tld);
 	  if (debug)
 	    printf ("FATAL\n");
 	  continue;
@@ -226,7 +224,7 @@ doit (void)
       rc = tld_check_4t (tld[i].in, tld[i].inlen, &errpos, tldtable);
       if (rc != tld[i].rc)
 	{
-	  fail ("TLD entry %ld failed: %d\n", i, rc);
+	  fail ("TLD entry %u failed: %d\n", i, rc);
 	  if (debug)
 	    printf ("FATAL\n");
 	  continue;
@@ -237,7 +235,7 @@ doit (void)
 
       if (rc != tld[i].rc)
 	{
-	  fail ("TLD entry %ld failed\n", i);
+	  fail ("TLD entry %u failed\n", i);
 	  if (debug)
 	    printf ("ERROR\n");
 	}
@@ -245,12 +243,12 @@ doit (void)
 	{
 	  if (debug)
 	    printf ("returned errpos %ld expected errpos %ld\n",
-		    errpos, tld[i].errpos);
+		    (long) errpos, (long) tld[i].errpos);
 
 	  if (tld[i].errpos != errpos)
 	    {
-	      fail ("TLD entry %ld failed because errpos %ld != %ld\n", i,
-		    tld[i].errpos, errpos);
+	      fail ("TLD entry %u failed because errpos %ld != %ld\n", i,
+		    (long) tld[i].errpos, (long) errpos);
 	      if (debug)
 		printf ("ERROR\n");
 	    }
@@ -262,12 +260,12 @@ doit (void)
 	rc = tld_check_8z (tld[i].example, &errpos, NULL);
 	if (rc != tld[i].rc)
 	  {
-	    fail ("TLD entry %ld failed\n", i);
+	    fail ("TLD entry %u failed\n", i);
 	    if (debug)
 	      printf ("ERROR\n");
 	  }
 	if (debug)
-	  printf ("TLD entry %ld tld_check_8z (%s)\n", i, tld[i].example);
+	  printf ("TLD entry %u tld_check_8z (%s)\n", i, tld[i].example);
       }
     }
 }
